jupyter_notebook/requirements.txt
python-dotenv==1.0.1
jupyterlab==4.0.7
ipywidgets==8.1.1
requests==2.32.3
numpy==1.26.4
pandas==2.1.1
scipy==1.11.3
scikit-learn==1.3.1
matplotlib==3.8.0
gensim==4.3.3
torch==2.6.0
transformers==4.48.3
tqdm==4.67.1
openai==1.61.1
gradio==5.15.0
langchain==0.3.21
tiktoken==0.8.0
faiss-cpu==1.10.0
langchain-openai==0.3.4
langchain-experimental==0.3.4
langchain-chroma==0.2.1
datasets==3.2.0
sentencepiece==0.2.0
#google-generativeai==0.8.4
anthropic==0.45.2
unstructured==0.16.20
chromadb==0.6.3
plotly==6.0.0
jupyter-dash==0.4.2
beautifulsoup4==4.12.2
pydub==0.25.1
modal==0.73.23
accelerate==1.3.0
bitsandbytes==0.45.2
psutil==5.9.5
setuptools==68.2.2
speedtest-cli==2.1.3
sentence-transformers==3.4.1
feedparser==6.0.11
psycopg2-binary==2.9.10
pymongo==4.6.1
motor==3.3.2  # If you need async support
dnspython==2.5.0  # Helpful for MongoDB connection strings
ijson== 3.3.0
langgraph==0.3.18
langchain-google-genai==2.1.0
langchain-community==0.3.20
duckduckgo-search==7.5.2
langchain-google-vertexai==2.0.15
google-genai==1.7.0
langchain-core==0.3.45
pygraphviz==1.14
pgeocode==0.5.0